+++ src/DlOperandTable.h
@@ -34,6 +34,16 @@ using ImmOp = int64_t;
 
 using RegOp = std::string;
 
+struct PrefetchOp
+{
+    std::string value;
+};
+
+struct BarrierOp
+{
+    std::string value;
+};
+
 struct IndirectOp
 {
     std::string reg1;
@@ -44,13 +54,19 @@ struct IndirectOp
     int size;
 };
 
-constexpr bool operator<(const IndirectOp &LHS, const IndirectOp &RHS) noexcept;
-souffle::tuple &operator<<(souffle::tuple &t, const IndirectOp &op);
+constexpr bool operator<(const IndirectOp& LHS, const IndirectOp& RHS) noexcept;
+souffle::tuple& operator<<(souffle::tuple& t, const IndirectOp& op);
+
+bool operator<(const PrefetchOp& LHS, const PrefetchOp& RHS) noexcept;
+souffle::tuple& operator<<(souffle::tuple& t, const PrefetchOp& op);
+
+bool operator<(const BarrierOp& LHS, const BarrierOp& RHS) noexcept;
+souffle::tuple& operator<<(souffle::tuple& t, const BarrierOp& op);
 
 template <class T>
-souffle::tuple &operator<<(souffle::tuple &t, const std::pair<T, uint64_t> &pair)
+souffle::tuple& operator<<(souffle::tuple& t, const std::pair<T, uint64_t>& pair)
 {
-    auto &[elem, id] = pair;
+    auto& [elem, id] = pair;
     t << id << elem;
     return t;
 }
@@ -61,13 +77,16 @@ private:
     // we reserve 0 for empty operators
     uint64_t curr_index = 1;
     template <typename T>
-    int64_t addToTable(std::map<T, uint64_t> &opTable, T op);
+    int64_t addToTable(std::map<T, uint64_t>& opTable, T op);
 
 public:
     std::map<ImmOp, uint64_t> immTable;
     std::map<RegOp, uint64_t> regTable;
     std::map<IndirectOp, uint64_t> indirectTable;
+    std::map<PrefetchOp, uint64_t> prefetchTable;
+    std::map<BarrierOp, uint64_t> barrierTable;
     int64_t add(std::variant<ImmOp, RegOp, IndirectOp> op);
+    int64_t add(std::variant<ImmOp, RegOp, IndirectOp, PrefetchOp, BarrierOp> op);
 };
 
 #endif /* SRC_DL_OPERATOR_TABLE_H_ */